Eliminate `off_t`.

+++ Storage/Disk/DiskImage/Formats/Utility/ImplicitSectors.cpp
@@ -21,8 +21,8 @@ using namespace Storage::Disk;
 std::shared_ptr<Track> Storage::Disk::track_for_sectors(const uint8_t *const source, int number_of_sectors, uint8_t track, uint8_t side, uint8_t first_sector, uint8_t size, bool is_double_density) {
 	std::vector<Storage::Encodings::MFM::Sector> sectors;
 
-	off_t byte_size = off_t(128 << size);
-	off_t source_pointer = 0;
+	size_t byte_size = size_t(128 << size);
+	size_t source_pointer = 0;
 	for(int sector = 0; sector < number_of_sectors; sector++) {
 		sectors.emplace_back();
